Feature/2021.2 (#24)
* Remove pluginUntilBuild

I'm not going to keep updating this for forever.
Instead, if the plugin does not work in a newer IntelliJ I'll look into it then.
Added verification for 2021.2.
Note: Trying to not bump version, since the GitHub Action should do this for me.

* Test build: Downgrade ktlint

* Force update build

Copy/paste from template

* Update pluginUntilBuild

Field is mandatory for verification. Ugh. Trying *

* Update pluginUntilBuild

* failed, trying something random but large

* Update pluginUntilBuild

9001.* failed, trying something random but large

* Update build from template

* Test build

* Update build java version when run github action

* Add latest to plugin verifier ide version

* Update pluginVerifierIdeVersions

Change from latest to 2021.2, seems "latest" didn't work - at least not in GitHub Actions :shrug:

Co-authored-by: KylinDC <inlovingdays@gmail.com>