Add support for AVA4 (#55)
* Add support for AVA4

The cli.js-file have been moved to lib-folder, but it doesn't work as before.
I'm not 100% sure why the file was targeted instead of the .bin-one,
could have been as simple as not having to deal with .cmd in windows.
Anyway, here comes a new solution; pick the first of these:
  1. node_modules/ava/cli.js
  2. node_modules/.bin/ava.cmd
  3. node_modules/.bin/ava

I haven't tested the .cmd one, as I don't have Windows here at the moment.
Also had to tweak the settings, hope it doesn't break.

* Update cli path calculation

Either use entrypoints/cli.mjs for >=4 or ava/cli.js for <4.

* Bumping plugin version

IIRC this was a manual step. Bumping minor.
This might even be a major, but hoping minor will do.